refactor code to separate files
